#23 refactor: 메인 페이지의 뷰 컴포넌트를 별도 파일로 분리, 기존 global.d.ts로 관리하던 타입 위치 변경 및 의존성 수정
