+++ pages/index.tsx
@@ -1,24 +1,33 @@
 // FIXME: 즐겨찾기, 출처별 필터 등 필터링 기능 작동 안함
-// TODO: 이 컴포넌트는 데이터 가공용으로 사용, 뷰는 별도 컴포넌트로 분리
 // TODO: 페이지 네비게이터는 5개까지 + 맨 처음, 맨 마지막 페이지만 표시, 그 외는 줄임표로
-import React from "react";
-import Search from "components/search";
+import { useEffect, useState } from "react";
 import RequestControllers from "controllers";
-import Card from "components/card";
-import { ParsedFeedsDataType } from "types/global";
-import { handleSort } from "common/helpers";
-import { SORT_STANDARD, SORT_STANDARD_STATE } from "common/constants";
-import SelectBox from "components/common/SelectBox";
 import { AxiosResponse } from "axios";
-import Modal from "components/modal";
-import SubscriptionDialogBox from "components/feeds";
-import SubscribeNew from "components/feeds/SubscribeNew";
-import CancelSubscription from "components/feeds/CancelSubscription";
 import { useQuery, useInfiniteQuery } from "@tanstack/react-query";
-import FilterBySource from "components/feeds/FilterBySource";
 import useFilters from "hooks/useFilters";
-import FilterByText from "components/feeds/FilterByText";
 import { SEARCH_OPTIONS } from "components/feeds/FilterByText";
+import MainPage from "components/main";
+
+export interface ParsedFeedsDataType {
+    id: string;
+    title: string | null;
+    description: string | null;
+    link: string | null;
+    pubDate: string | null;
+    origin: string | null;
+    isRead: boolean | null;
+    isFavorite: boolean | null;
+    [key: string]: number | string | boolean | null;
+}
+
+export interface ParseResultType {
+    id: number;
+    originName: string | null;
+    originLink: string | null;
+    lastFeedsLength: number;
+    latestFeedTitle: string | null;
+    feeds?: ParsedFeedsDataType[];
+}
 
 interface IndexProps {
     feeds: string;
@@ -30,29 +39,17 @@ interface RenewedFeedsData {
     count: number;
 }
 
-type ModalKeys = "addSubscription" | "cancelSubscription" | "filterBySource";
-
-type ModalStateType = {
-    [key in ModalKeys]: boolean;
-};
-
 export default function Index({ feeds, sources }: IndexProps) {
     const { getDataFrom } = new RequestControllers();
-    const [currentSort, setCurrentSort] = React.useState(0);
-    const [modalState, setModalState] = React.useState<ModalStateType>({
-        addSubscription: false,
-        cancelSubscription: false,
-        filterBySource: false,
-    });
+    const [currentSort, setCurrentSort] = useState(0);
+    const [isFilterFavorite, setIsFilterFavorite] = useState<boolean>(false);
     const [observerElement, setObserverElement] =
-        React.useState<HTMLDivElement | null>(null);
-    const [isFilterFavorite, setIsFilterFavorite] =
-        React.useState<boolean>(false);
-    const [newFeeds, setNewFeeds] = React.useState<ParsedFeedsDataType[]>([]);
-    const [totalCount, setTotalCount] = React.useState<number>(0);
-    const [isMobileLayout, setIsMobileLayout] = React.useState<boolean>(false);
-    const [currentPage, setCurrentPage] = React.useState<number>(1);
-    const [formerFeedsList, setFormerFeedsList] = React.useState<any>({});
+        useState<HTMLDivElement | null>(null);
+    const [newFeeds, setNewFeeds] = useState<ParsedFeedsDataType[]>([]);
+    const [totalCount, setTotalCount] = useState<number>(0);
+    const [isMobileLayout, setIsMobileLayout] = useState<boolean>(false);
+    const [currentPage, setCurrentPage] = useState<number>(1);
+    const [formerFeedsList, setFormerFeedsList] = useState<any>({});
     const [sourceDisplayState, setSourceDisplayState] = useFilters(
         sources,
         true
@@ -61,7 +58,6 @@ export default function Index({ feeds, sources }: IndexProps) {
         JSON.stringify(Object.values(SEARCH_OPTIONS)),
         ""
     );
-    const startPageRef = React.useRef<HTMLElement | null>(null);
     const newFeedsRequestResult = useQuery<AxiosResponse<RenewedFeedsData>>(
         ["/feeds/new"],
         () => getDataFrom("/feeds/new")
@@ -107,25 +103,19 @@ export default function Index({ feeds, sources }: IndexProps) {
         }
     };
 
-    const handleClick = (target: ModalKeys) => () => {
-        document.documentElement.scrollTo({ top: 0 });
-        closeModal(target, !modalState[target])();
-    };
-
-    const closeModal =
-        (target: ModalKeys, lastModalState = false) =>
-        () => {
-            setModalState((modalStateObject) => ({
-                ...modalStateObject,
-                [target]: lastModalState,
-            }));
-        };
-
     const filterFavorites = () => {
         setIsFilterFavorite(!isFilterFavorite);
     };
 
-    React.useEffect(() => {
+    const updateObserverElement = (element: HTMLDivElement) => {
+        setObserverElement(element);
+    };
+
+    const updateCurrentPage = (value: number | ((value: number) => number)) => {
+        setCurrentPage(value);
+    };
+
+    useEffect(() => {
         if (typeof window !== "undefined") {
             const callback = () => {
                 if (window.innerWidth > 768) {
@@ -140,7 +130,7 @@ export default function Index({ feeds, sources }: IndexProps) {
         }
     }, []);
 
-    React.useEffect(() => {
+    useEffect(() => {
         if (feeds) {
             const {
                 data,
@@ -160,7 +150,7 @@ export default function Index({ feeds, sources }: IndexProps) {
         }
     }, [feeds]);
 
-    React.useEffect(() => {
+    useEffect(() => {
         if (storedFeed && storedFeed.pages) {
             const dataArray = JSON.parse(
                 storedFeed.pages[storedFeed.pages.length - 1].data
@@ -171,7 +161,7 @@ export default function Index({ feeds, sources }: IndexProps) {
         }
     }, [storedFeed, isMobileLayout]);
 
-    React.useEffect(() => {
+    useEffect(() => {
         if (
             newFeedsRequestResult != null &&
             typeof newFeedsRequestResult !== "string"
@@ -181,23 +171,17 @@ export default function Index({ feeds, sources }: IndexProps) {
         }
     }, [newFeedsRequestResult]);
 
-    React.useEffect(() => {
-        if (modalState.addSubscription || modalState.cancelSubscription) {
-            document.documentElement.style.overflow = "hidden";
-        } else {
-            document.documentElement.style.overflow = "auto";
-        }
-    }, [modalState, startPageRef]);
-
-    React.useEffect(() => {
+    useEffect(() => {
         if (!isMobileLayout) {
             refetchStoredFeeds();
         }
     }, [isFilterFavorite, searchTexts, currentPage, isMobileLayout]);
 
-    React.useEffect(() => {
+    useEffect(() => {
         if (isMobileLayout) {
-            const firstEmptyPageIndex = (Object.values(formerFeedsList) as any[]).findIndex((value: any[]) => value.length === 0);
+            const firstEmptyPageIndex = (
+                Object.values(formerFeedsList) as any[]
+            ).findIndex((value: any[]) => value.length === 0);
             setCurrentPage(firstEmptyPageIndex);
             Object.keys(formerFeedsList).forEach((key: string, index) => {
                 if (index > firstEmptyPageIndex) {
@@ -208,13 +192,21 @@ export default function Index({ feeds, sources }: IndexProps) {
                 }
             });
         } else {
-            const fetchedPages = (Object.values(formerFeedsList) as any[]).reduce((totalNumber: number, currentDataArray: any[]) => currentDataArray.length > 0 ? totalNumber += 1 : totalNumber, 0);
+            const fetchedPages = (
+                Object.values(formerFeedsList) as any[]
+            ).reduce(
+                (totalNumber: number, currentDataArray: any[]) =>
+                    currentDataArray.length > 0
+                        ? (totalNumber += 1)
+                        : totalNumber,
+                0
+            );
             setCurrentPage(fetchedPages > 0 ? fetchedPages : 1);
         }
     }, [isMobileLayout]);
 
-    React.useEffect(() => {
-        const doh = newFeedsRequestResult
+    useEffect(() => {
+        const fetchedNewFeeds = newFeedsRequestResult
             ? newFeedsRequestResult?.data
             : newFeeds;
         setFormerFeedsList((previousObject: any) => ({
@@ -223,12 +215,12 @@ export default function Index({ feeds, sources }: IndexProps) {
                 previousObject[currentPage].length === 0
                     ? previousObject[currentPage]
                           ?.slice(previousObject[currentPage].length)
-                          .concat(doh)
+                          .concat(fetchedNewFeeds)
                     : previousObject[currentPage],
         }));
     }, [newFeedsRequestResult, newFeeds]);
 
-    React.useEffect(() => {
+    useEffect(() => {
         if (typeof window !== "undefined" && observerElement != null) {
             const observerOption: IntersectionObserverInit = {
                 threshold: 0.5,
@@ -256,158 +248,24 @@ export default function Index({ feeds, sources }: IndexProps) {
         }
     }, [observerElement, hasNextPage]);
 
-    const feedsToDisplay = feedsFromServer
-        ? feedsFromServer
-              ?.sort(
-                  handleSort(
-                      SORT_STANDARD_STATE[currentSort],
-                      checkShouldSortByReverse(currentSort)
-                  )
-              )
-              .map((feed: ParsedFeedsDataType) => (
-                  <Card
-                      cardData={feed}
-                      key={feed.id}
-                      refetchFeeds={refetchStoredFeeds}
-                  />
-              ))
-        : [];
-
-    const pageIndicator = Array.from(
-        { length: Math.ceil(totalCount / 10) },
-        (v, k) => k + 1
-    ).map((pageIndex: number) => (
-        <li
-            key={`page_${pageIndex}`}
-            className="list-none"
-            onClick={() => {
-                setCurrentPage(pageIndex);
-            }}
-        >
-            <button
-                className={`${
-                    currentPage === pageIndex ? "text-blue-500 font-bold" : ""
-                }`}
-            >
-                {pageIndex}
-            </button>
-        </li>
-    ));
-
     return (
-        <article
-            className="flex-center flex-col w-full h-max min-h-full bg-neutral-100 dark:bg-neutral-800 dark:text-neutral-200"
-            ref={startPageRef}
-        >
-            <section className="flex-center w-full h-1/3 my-32 md:w-[768px]">
-                <Search />
-            </section>
-            <section className="flex flex-col items-center w-full h-max md:w-[768px]">
-                <section>
-                    <section className="flex justify-between h-8 mb-4">
-                        <section>
-                            <button
-                                className="mr-3 px-3 py-2 rounded-md shadow-md bg-neutral-100 text-xs text-neutral-700 dark:shadow-zinc-600 dark:bg-neutral-700 dark:text-neutral-200"
-                                onClick={handleClick("addSubscription")}
-                            >
-                                구독 추가
-                            </button>
-                            <button
-                                className="mr-3 px-3 py-2 rounded-md shadow-md bg-neutral-100 text-xs text-neutral-700 dark:shadow-zinc-600 dark:bg-neutral-700 dark:text-neutral-200"
-                                onClick={handleClick("cancelSubscription")}
-                            >
-                                구독 취소
-                            </button>
-                            <button
-                                className="mr-3 px-3 py-2 rounded-md shadow-md bg-neutral-100 text-xs text-neutral-700 dark:shadow-zinc-600 dark:bg-neutral-700 dark:text-neutral-200"
-                                onClick={filterFavorites}
-                            >
-                                즐겨찾기
-                            </button>
-                            <button
-                                className="mr-3 px-3 py-2 rounded-md shadow-md bg-neutral-100 text-xs text-neutral-700 dark:shadow-zinc-600 dark:bg-neutral-700 dark:text-neutral-200"
-                                onClick={handleClick("filterBySource")}
-                            >
-                                출처별 필터
-                            </button>
-                        </section>
-                        <FilterByText setTextFilter={setSearchTexts} />
-                        <SelectBox
-                            optionValues={SORT_STANDARD}
-                            customStyles="rounded-md shadow-md text-xs dark:shadow-zinc-600"
-                            setSortState={setSortState}
-                        />
-                    </section>
-                </section>
-                <section>{feedsToDisplay}</section>
-                {isMobileLayout ? (
-                    <div
-                        ref={setObserverElement}
-                        className="w-full h-[150px]"
-                    />
-                ) : (
-                    <ul className="flex justify-evenly items-center w-1/2 mt-10 mb-20">
-                        <button
-                            onClick={() => {
-                                if (currentPage !== 1) {
-                                    setCurrentPage(
-                                        (previousValue) => previousValue - 1
-                                    );
-                                }
-                            }}
-                        >
-                            &lt;
-                        </button>
-                        {pageIndicator}
-                        <button
-                            onClick={() => {
-                                if (
-                                    Math.ceil(totalCount / 10) !== currentPage
-                                ) {
-                                    setCurrentPage(
-                                        (previousValue) => previousValue + 1
-                                    );
-                                }
-                            }}
-                        >
-                            &gt;
-                        </button>
-                    </ul>
-                )}
-            </section>
-            {modalState.addSubscription && (
-                <Modal closeModal={closeModal("addSubscription")}>
-                    <SubscriptionDialogBox
-                        closeModal={closeModal("addSubscription")}
-                    >
-                        <SubscribeNew />
-                    </SubscriptionDialogBox>
-                </Modal>
-            )}
-            {modalState.cancelSubscription && (
-                <Modal closeModal={closeModal("cancelSubscription")}>
-                    <SubscriptionDialogBox
-                        closeModal={closeModal("cancelSubscription")}
-                    >
-                        <CancelSubscription sources={sources} />
-                    </SubscriptionDialogBox>
-                </Modal>
-            )}
-            {modalState.filterBySource && (
-                <Modal closeModal={closeModal("filterBySource")}>
-                    <SubscriptionDialogBox
-                        closeModal={closeModal("filterBySource")}
-                    >
-                        <FilterBySource
-                            displayState={sourceDisplayState}
-                            setDisplayFlag={setSourceDisplayState}
-                            closeModal={closeModal("filterBySource")}
-                            refetchFeeds={refetchStoredFeeds}
-                        />
-                    </SubscriptionDialogBox>
-                </Modal>
-            )}
-        </article>
+        <MainPage
+            feedsFromServer={feedsFromServer}
+            currentPage={currentPage}
+            setCurrentPage={updateCurrentPage}
+            currentSort={currentSort}
+            checkShouldSortByReverse={checkShouldSortByReverse}
+            setSortState={setSortState}
+            totalCount={totalCount}
+            isMobileLayout={isMobileLayout}
+            sources={sources}
+            sourceDisplayState={sourceDisplayState}
+            setSourceDisplayState={setSourceDisplayState}
+            updateObserverElement={updateObserverElement}
+            refetchStoredFeeds={refetchStoredFeeds}
+            setSearchTexts={setSearchTexts}
+            filterFavorites={filterFavorites}
+        />
     );
 }
