--- types/global.d.ts
@@ -1,21 +1,0 @@
-// TODO: index 페이지에서 선언하도록 수정
-export interface ParsedFeedsDataType {
-    id: string;
-    title: string | null;
-    description: string | null;
-    link: string | null;
-    pubDate: string | null;
-    origin: string | null;
-    isRead: boolean | null;
-    isFavorite: boolean | null;
-    [key: string]: number | string | boolean | null;
-}
-
-export interface ParseResultType {
-    id: number;
-    originName: string | null;
-    originLink: string | null;
-    lastFeedsLength: number;
-    latestFeedTitle: string | null;
-    feeds?: ParsedFeedsDataType[];
-}
